Committing changes from last night

+++ Robotics/Drivers/Hermes/Firmware/hermes/HMC5883L.h
@@ -66,7 +66,7 @@ class HMC5883L
 	  char* GetErrorText(int errorCode);
 
 	protected:
-	  void Write(int address, int byte);
+	  void Write(uint8_t address, uint8_t byte);
 	  uint8_t* Read(int address, int length);
 
 	private:
